+++ src/ZendeskApi.Client/Models/AuditEvents/CommentRedactionEvent.cs
@@ -5,6 +5,6 @@ namespace ZendeskApi.Client.Models.AuditEvents
     public class CommentRedactionEvent : AuditEvent, IAuditEvent
     {
         [JsonProperty("comment_id")]
-        public int CommentId { get; set; }
+        public long CommentId { get; set; }
     }
 }
